reorganise installables tests
