+++ tests/core/installables/test_hf_model.py
